feat: implementa gestao de mesas com geracao de QR Code e upload para S3
- Cria a entidade RestaurantTable e os respectivos DTOs para encapsulamento.
- Implementa o RestaurantTableRepository com verificacao de duplicidade (existsByTableNumber).
- Integra o QrCodeGenerateService para gerar o QR Code (via ZXing) e acionar a StoragePort para upload na AWS S3.
- Implementa o RestaurantTableService orquestrando a geracao da URL do front-end, processamento do S3 e persistencia da mesa.
- Cria o RestaurantTableController com endpoints REST e habilitacao de @CrossOrigin para consumo pelo front-end em Angular.

+++ src/main/java/com/gabrielmonteiro/qrcode/generator/entities/Product.java
@@ -5,7 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.util.function.LongFunction;
+import java.math.BigDecimal;
+
 
 @Entity
 @Table (name = "tbl_products")
@@ -18,13 +19,13 @@ public class Product {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
-    private Double price;
+    private BigDecimal price;
     private String description;
     private Boolean available;
     private String category;
     private String imageUrl;
 
-    public Product(String name, Double price, String description, Boolean available, String category, String imageUrl) {
+    public Product(String name, BigDecimal price, String description, Boolean available, String category, String imageUrl) {
         this.name = name;
         this.price = price;
         this.description = description;
@@ -32,5 +33,7 @@ public class Product {
         this.category = category;
         this.imageUrl = imageUrl;
     }
+
+
 
 }
